Move the initialization of `vehicle_ordering` back to `as_driving_graph`
